完整功能更新：后端API优化和前端功能增强
后端改进:
- UserController: 添加用户列表API,支持分页和财富等级过滤
- AuthService: 优化用户注册和登录逻辑
- JPushService: 完善多设备推送通知系统
- CallService/MessageService: 改进通话和消息服务
- 新增支付系统(PaymentController, AlipayService)
- 新增个人资料管理(ProfileController, ProfileService)
- 新增道具商城(PropMallController, PropMallService)
- 新增财富等级系统(WealthController, WealthService)
- 新增设备管理(DeviceController, UserDeviceService)
- 完善数据库初始化脚本

前端改进:
- 优化AndroidManifest.xml配置
- 完善用户详情页(UserDetailActivity)
- 改进个人资料编辑(MyProfileEditActivity)
- 优化充值功能(RechargeActivity, RechargeViewModel)
- 完善个人资料Compose界面(ProfileScreen, ProfileComposeHost)
- 优化道具商城界面(PropMallScreen, PropPurchaseScreen)
- 新增价格设置功能(PriceSettingsActivity)
- 新增设备管理功能(DeviceManagementActivity)
- 完善网络服务层(ApiService, NetworkService)
- 优化支付管理(PaymentManager, AlipayManager)
- 新增多个DTO和Model类支持新功能
- 完善ViewModel层(ProfileViewModel, PropPurchaseViewModel等)

技术文档:
- 添加API文档(个人资料、道具商城、支付系统)
- 添加系统架构文档(多设备推送、头像上传)
- 添加测试脚本和配置脚本

数据库:
- 新增多个数据表初始化脚本
- 完善用户、钱包、订单等表结构

+++ app/src/main/java/com/example/myapplication/viewmodel/PropPurchaseViewModel.kt
@@ -27,9 +27,9 @@ class PropPurchaseViewModel : ViewModel() {
         _uiState.value = _uiState.value.copy(
             luckyNumber = luckyNumber,
             token = token,
-            originalPrice = luckyNumber.price,
-            discountAmount = calculateDiscount(luckyNumber.price),
-            finalPrice = calculateFinalPrice(luckyNumber.price)
+            originalPrice = luckyNumber.price.toLong(),
+            discountAmount = calculateDiscount(luckyNumber.price.toLong()),
+            finalPrice = calculateFinalPrice(luckyNumber.price.toLong())
         )
     }
